feat: hooks are now deep imports
