+++ examples/basic/src/app.tsx
@@ -1,4 +1,4 @@
-import { useAsset } from "ultra/client.js";
+import useAsset from "ultra/hooks/use-asset.js";
 
 export default function App() {
   console.log("Hello world!");
