+++ hooks/flush-effect-context.js
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+/**
+ * @type {React.Context<null | (handler: () => React.ReactNode>)}
+ */
+const FlushEffectsContext = createContext(
+  null,
+);
+
+export default FlushEffectsContext;
