+++ hooks/use-flush-effects.js
@@ -1,17 +1,11 @@
-import { createContext, useContext } from "react";
-
-/**
- * @type {React.Context<null | (handler: () => React.ReactNode>)}
- */
-export const FlushEffectsContext = createContext(
-  null,
-);
+import { useContext } from "react";
+import FlushEffectsContext from "./flush-effect-context.js";
 
 /**
  * @param {() => React.ReactNode} callback
  * @returns {void}
  */
-export function useFlushEffects(callback) {
+export default function useFlushEffects(callback) {
   const addFlushEffects = useContext(FlushEffectsContext);
   // Should have no effects on client where there's no flush effects provider
   if (addFlushEffects) {
